# PERF: use buffer pool in json formatter