@taylorchu 2018-05-17
@sirupsen this idea is also in text_formatter. Could you take a look? thanks!